+++ src/main/java/com/vonage/client/auth/RequestSigning.java
@@ -25,6 +25,7 @@ import org.apache.http.NameValuePair;
 import org.apache.http.message.BasicNameValuePair;
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
+import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.time.Instant;
@@ -154,9 +155,12 @@ public class RequestSigning {
      * @param secretKey The pre-shared secret key used by the sender of the request to create the signature.
      *
      * @return true if the signature is correct for this request and secret key.
+     *
+     * @deprecated This method will be removed in a future release due to the dependency on javax.servlet.
      */
+    @Deprecated
     public static boolean verifyRequestSignature(HttpServletRequest request, String secretKey) {
-        return verifyRequestSignature(request, secretKey, System.currentTimeMillis());
+        return verifyRequestSignature(request, secretKey, HashUtil.HashType.MD5);
     }
 
     /**
@@ -167,48 +171,70 @@ public class RequestSigning {
      * @param hashType Hash type to be used to construct request parameters.
      *
      * @return true if the signature is correct for this request and secret key.
+     *
+     * @deprecated This method will be removed in a future release due to the dependency on javax.servlet.
      */
+    @Deprecated
     public static boolean verifyRequestSignature(HttpServletRequest request, String secretKey, HashUtil.HashType hashType) {
-        return verifyRequestSignature(request, secretKey, System.currentTimeMillis(), hashType);
+        try {
+            return verifyRequestSignature(
+                    request.getContentType(),
+                    request.getInputStream(),
+                    request.getParameterMap(),
+                    secretKey, System.currentTimeMillis(), hashType
+            );
+        }
+        catch (IOException ex) {
+            throw new VonageUnexpectedException("Error encountered when opening input stream for request", ex);
+        }
     }
 
     /**
-     * Verifies the signature in an HttpServletRequest.
-     * Hashing strategy is MD5.
+     * Verifies the signature in an HttpServletRequest. Hashing strategy is MD5.
      *
-     * @param request The HttpServletRequest to be verified.
+     * @param contentType The request Content-Type header.
+     * @param inputStream The request data stream.
+     * @param parameterMap The request parameters.
      * @param secretKey The pre-shared secret key used by the sender of the request to create the signature.
      * @param currentTimeMillis The current time, in milliseconds.
      *
      * @return true if the signature is correct for this request and secret key.
      */
-     protected static boolean verifyRequestSignature(HttpServletRequest request,
+     protected static boolean verifyRequestSignature(String contentType,
+                                                     InputStream inputStream,
+                                                     Map<String, String[]> parameterMap,
                                                      String secretKey,
                                                      long currentTimeMillis) {
-        return verifyRequestSignature(request, secretKey, currentTimeMillis, HashUtil.HashType.MD5);
+        return verifyRequestSignature(contentType, inputStream, parameterMap,
+                secretKey, currentTimeMillis, HashUtil.HashType.MD5
+        );
     }
 
     /**
      * Verifies the signature in an HttpServletRequest.
      *
-     * @param request The HttpServletRequest to be verified.
+     * @param contentType The request Content-Type header.
+     * @param inputStream The request data stream.
+     * @param parameterMap The request parameters.
      * @param secretKey The pre-shared secret key used by the sender of the request to create the signature.
      * @param currentTimeMillis The current time, in milliseconds.
      * @param hashType Hash type to be used to construct request parameters.
      *
      * @return true if the signature is correct for this request and secret key.
      */
-    protected static boolean verifyRequestSignature(HttpServletRequest request,
+    protected static boolean verifyRequestSignature(String contentType,
+                                                    InputStream inputStream,
+                                                    Map<String, String[]> parameterMap,
                                                     String secretKey,
                                                     long currentTimeMillis,
                                                     HashUtil.HashType hashType) {
 
         // Construct a sorted list of the name-value pair parameters supplied in the request, excluding the signature parameter
         Map<String, String> sortedParams = new TreeMap<>();
-        if (request.getContentType() != null && request.getContentType().equals(APPLICATION_JSON)) {
+        if (APPLICATION_JSON.equals(contentType) && inputStream != null) {
             ObjectMapper mapper = new ObjectMapper();
-            try{
-                Map<String,String> params = mapper.readValue(request.getInputStream(), new TypeReference<Map<String,String>>(){});
+            try {
+                Map<String,String> params = mapper.readValue(inputStream, new TypeReference<Map<String,String>>(){});
                 for (Map.Entry<String, String> entry : params.entrySet()) {
                     String name = entry.getKey();
                     String value = entry.getValue();
@@ -224,11 +250,11 @@ public class RequestSigning {
             }
         }
         else {
-            for (Map.Entry<String, String[]> entry: request.getParameterMap().entrySet()) {
+            for (Map.Entry<String, String[]> entry : parameterMap.entrySet()) {
                 String name = entry.getKey();
                 String value = entry.getValue()[0];
                 log.info(name + " = " + value);
-                if (value == null || value.trim().equals("")) {
+                if (value == null || value.trim().isEmpty()) {
                     continue;
                 }
                 sortedParams.put(name, value);